# frontend/public/openapi.yaml
openapi: 3.0.3
info:
  title: Trade Nexus API
  version: 1.0.0
  description: |
    Comprehensive API for Trade Nexus - CFTC Position Limits Compliance Platform

    ## Overview
    Trade Nexus provides a complete suite of APIs for managing commodity trading position limits,
    risk management, compliance reporting, and regulatory submissions to the CFTC.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain a token via `/api/auth/login`.

    ## Rate Limiting
    - Standard users: 100 requests per minute
    - API keys: 1000 requests per minute

    ## Base URLs
    - **Production**: https://trade-nexus-api.metabilityllc1.workers.dev
    - **Development**: https://trade-nexus-api-dev.metabilityllc1.workers.dev

  contact:
    name: Trade Nexus Support
    email: support@tradenexus.com
  license:
    name: Proprietary
servers:
  - url: https://trade-nexus-api-dev.metabilityllc1.workers.dev
    description: Development server
  - url: https://trade-nexus-api.metabilityllc1.workers.dev
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        category:
          type: string
          enum: [limit, compliance, system]
        unread:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - severity
security:
  - BearerAuth: []
paths:
...
  /api/alerts:
    get:
      tags:
        - Alerts
      summary: List alerts
      description: Retrieve all alerts for the authenticated user
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
          description: Filter by read/unread status
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
          description: Filter by severity
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Maximum number of alerts to return
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
  /api/alerts/{id}/read:
    put:
      tags:
        - Alerts
      summary: Mark alert as read
      description: Mark a specific alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alert marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...